requirements.txt
asn1crypto==1.0.0
astroid==1.6.6
autopep8==1.4.4
backports.functools-lru-cache==1.5
cachetools==3.1.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
configparser==3.7.4
cryptography==2.7
enum34==1.1.6
futures==3.2.0
google-api-python-client==1.7.9
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.0
httplib2==0.13.0
idna==2.8
ipaddress==1.0.22
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.2
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
pycparser==2.19
pylint==1.9.4
pyOpenSSL==19.0.0
requests==2.22.0
requests-oauthlib==1.2.0
rsa==4.0
singledispatch==3.4.0.3
six==1.12.0
slackclient==1.3.2
uritemplate==3.0.0
urllib3==1.25.3
usabilla-api==2.0.1
websocket-client==0.54.0
wrapt==1.11.2
